requirements.txt
parsel
playwright
python-dotenv